refactor(layering): Phase 8 S3 — cassette staging leaves Course (refs #802)
Third step of the A1/A3 design. The HTTP-replay cassette staging
maintenance moves off the core Course class:

- workers/notebook/http_replay_cassette.py ->
  infrastructure/http_replay_mitm/http_replay_cassette.py (its deps were
  already downward-only: vcr_format + the core trace leaf)
- Course.merge_mitmproxy_cassette_staging and
  Course._sweep_orphan_cassette_staging_files become functions in
  infrastructure/http_replay_mitm/cassette_staging.py, taking the
  canonical paths from the new public Course.http_replay_canonical_paths()
- sweeping is the entry points' job now: clm build's pre-stage hook and
  watch mode's FileEventHandler call it; Course.process_file/process_all
  no longer sweep for themselves (also removes build.py's private-symbol
  cross-call, an A9 offender)

Ratchet: 5 -> 4 edges over 3 files (S4 identity reads + S5 voiceover
merge remain).
