+++ tests/cli/test_build_command.py
@@ -1937,7 +1937,7 @@ class TestProcessCourseInvokesCassetteSweep:
     """Regression for issue #145.
 
     The pre-build orphan staging-cassette sweep
-    (:meth:`Course._sweep_orphan_cassette_staging_files`) was documented
+    (now ``cassette_staging.sweep_orphan_cassette_staging_files``) was documented
     to run before every ``clm build`` but was actually only invoked from
     ``Course.process_all`` / ``Course.process_file``. The ``clm build``
     path goes through ``process_course_with_backend`` →
@@ -1953,10 +1953,10 @@ class TestProcessCourseInvokesCassetteSweep:
         from clm.cli.commands.build import process_course_with_backend
 
         source = inspect.getsource(process_course_with_backend)
-        assert "_sweep_orphan_cassette_staging_files" in source, (
+        assert "sweep_orphan_cassette_staging_files" in source, (
             "process_course_with_backend must invoke "
-            "course._sweep_orphan_cassette_staging_files() before the "
-            "stage loop (issue #145). If this assertion fires, the call "
+            "cassette_staging.sweep_orphan_cassette_staging_files() before "
+            "the stage loop (issue #145). If this assertion fires, the call "
             "was removed or moved out of the build entry path — restore "
             "it or the orphan cleanup stops happening during normal builds."
         )
